Remove json.h include from various headers
