+++ src/ascii_art.h
@@ -5,8 +5,9 @@
 #include <string>
 #include <vector>
 
-#include "json.h"
 #include "type_id.h"
+
+class JsonObject;
 
 
 class ascii_art
