+++ src/projectile.cpp
@@ -8,6 +8,7 @@
 #include "ammo_effect.h"
 #include "explosion.h"
 #include "game.h"
+#include "json.h"
 #include "item.h"
 #include "map.h"
 #include "map_iterator.h"
